Add simple workflow to check the output
So that pull requests can be merged with more confidence.

This also checks that `go mod tidy` doesn't result in changes, in case
`go.sum` is outdated. We could check explicitly for just the `git diff`
of `go.sum`, but there should be nothing else changing either in this
workflow, so have kept this a more generic check.